Controle por validação para adicionar e remover usuarios

+++ src/delete.php
@@ -1,5 +1,5 @@
 <?php
-if(isset($_GET["userDelete"])){
+if(isset($_GET["userDelete"]) && isset($_COOKIE["validado"]) && $_COOKIE["validado"] == $_COOKIE["PHPSESSID"]){
     $pdo = new PDO("mysql:host=localhost;dbname=adminsistema", "root", "toor"); 
     $consulta = $pdo->prepare('DELETE FROM admin WHERE id = :id');
     $consulta->bindParam(':id', $_GET["userDelete"]);
